Do not include processes that we can't parse stats_line for
In case reading /proc/<pid>/stats line fails for whatever reason
(short living process which dies between opendir and getline calls,
for example, or if operating system limits are pressured, so the
file can't be open), we will not try to calculate the CPU and memory
usage for this process, since we would work with garbage at that point.

If parseProcStats fails, it will now return false, and the process/thread
will not be marked, removing them in the GC cycle.